+++ vars/buildViaJervis.groovy
@@ -330,6 +330,7 @@ def buildViaJervis(String jervis_yaml, List folder_listing) {
         }
         List publishableItems = pipeline_generator.publishableItems
         if(publishableItems) {
+            //admin defining default settings for publishers they support
             pipeline_generator.collect_settings_defaults = [
                 artifacts: [
                     allowEmptyArchive: false,
@@ -340,12 +341,25 @@ def buildViaJervis(String jervis_yaml, List folder_listing) {
                 cobertura: [
                     autoUpdateHealth: false,
                     autoUpdateStability: false,
+                    failNoReports: false,
                     failUnhealthy: false,
                     failUnstable: false,
                     maxNumberOfBuilds: 0,
                     onlyStable: false,
                     sourceEncoding: 'ASCII',
-                    zoomCoverageChart: false
+                    zoomCoverageChart: false,
+                    methodCoverageTargets: '80, 0, 0',
+                    lineCoverageTargets: '80, 0, 0',
+                    conditionalCoverageTargets: '70, 0, 0'
+                ],
+                html: [
+                    allowMissing: false,
+                    alwaysLinkToLastBuild: false,
+                    includes: '**/*',
+                    keepAll: false,
+                    reportFiles: 'index.html',
+                    reportName: 'HTML Report',
+                    reportTitles: ''
                 ],
                 junit: [
                     allowEmptyResults: false,
@@ -353,7 +367,30 @@ def buildViaJervis(String jervis_yaml, List folder_listing) {
                     keepLongStdio: false
                 ]
             ]
-            pipeline_generator.collect_settings_filesets = [artifacts: ['excludes']]
+            //admin requiring stashes for HTML publisher to be formed a compatible way.
+            pipeline_generator.stashmap_preprocessor = [
+                html: { Map settings ->
+                    settings['includes']?.tokenize(',').collect {
+                        "${settings['path']  -~ '/$' -~ '^/'}/${it}"
+                    }.join(',').toString()
+                }
+            ]
+            //admin supporting optional list or string for filesets in default settings
+            pipeline_generator.collect_settings_filesets = [artifacts: ['excludes'], html: ['includes']]
+            //admin requiring regex validation of specific jenkins.collect setinggs
+            //if a user fails the input validation it falls back to the default option
+            //if an invalid path is specified for HTML publisher then do not attempt to collect
+            String cobertura_targets_regex = '([0-9]*\\.?[0-9]*,? *){3}[^,]$'
+            pipeline_generator.collect_settings_validation = [
+                cobertura: [
+                    methodCoverageTargets: cobertura_targets_regex,
+                    lineCoverageTargets: cobertura_targets_regex,
+                    conditionalCoverageTargets: cobertura_targets_regex
+                ],
+                html: [
+                    path: '''^[^,\\:*?"'<>|]+$'''
+                ]
+            ]
             stage("Publish results") {
                 for(String name : publishableItems) {
                     unstash name
@@ -369,18 +406,29 @@ def buildViaJervis(String jervis_yaml, List folder_listing) {
                                              caseSensitive: item['caseSensitive']
                             break
                         case 'cobertura':
-                            step([
-                                    $class: 'CoberturaPublisher',
+                            cobertura coberturaReportFile: item['path'],
                                     autoUpdateHealth: item['autoUpdateHealth'],
                                     autoUpdateStability: item['autoUpdateStability'],
-                                    coberturaReportFile: item['path'],
+                                    failNoReports: item['failNoReports'],
                                     failUnhealthy: item['failUnhealthy'],
                                     failUnstable: item['failUnstable'],
                                     maxNumberOfBuilds: item['maxNumberOfBuilds'],
                                     onlyStable: item['onlyStable'],
                                     sourceEncoding: item['sourceEncoding'],
-                                    zoomCoverageChart: item['zoomCoverageChart']
-                            ])
+                                    zoomCoverageChart: item['zoomCoverageChart'],
+                                    methodCoverageTargets: item['methodCoverageTargets'],
+                                    lineCoverageTargets: item['lineCoverageTargets'],
+                                    conditionalCoverageTargets: item['conditionalCoverageTargets']
+                            break
+                        case 'html':
+                            publishHTML allowMissing: item['allowMissing'],
+                                        alwaysLinkToLastBuild: item['alwaysLinkToLastBuild'],
+                                        includes: item['includes'],
+                                        keepAll: item['keepAll'],
+                                        reportDir: item['path'],
+                                        reportFiles: item['reportFiles'],
+                                        reportName: item['reportName'],
+                                        reportTitles: item['reportTitles']
                             break
                         case 'junit':
                             junit allowEmptyResults: item['allowEmptyResults'],